+++ conjugate_gradients.m
@@ -1,5 +1,9 @@
+%% DESCRIPTION
+
 % CONJUGATE GRADIENTS
 % Minimization of Unconstrained Quadratic problems
+
+%% Settings
 
 clc
 clear all
@@ -7,19 +11,35 @@ close all
 
 format long;
 
-% Generate new problem
-[n, Q, x, b, f, df, x_old] = quadratic_form();
+%% Problem
 
-% A matrix to store all 'x' vectors during iterations
-X = zeros(1, n);
-X(1, :) = x_old';
+% Call 'problem' function to generate a new problem
+% [n, f, df, A, x, b, X] = norm_problem();
+[n, f, df, Q, x, X] = quadratic_problem();
 
-% Hyper parameters
+%% Hyper parameters
+
+% Set a treshold to stop the loop
 treshold = 1e-6;
-alpha = 0.001;
 
-% Define an index for count the number of iterations 
-i = 1;
+% Learning rate
+alpha = 0.0001;
+
+% Armijo and Goldstein Rule parameters
+% alpha_max = randi([1, 4]);
+% alpha_max = 1;
+% c = 0.1;
+
+% Wolfe Conditions parameters
+% alpha_max = randi([1, 4]);
+% alpha_max = 1;
+% c1 = 0.1;
+% c2 = 0.2;
+
+%% Loop settings
+
+% Define an index for count the number of iterations
+k = 1;
 
 % Define a boolean flag variable to controle the while loop
 flag = true;
@@ -27,49 +47,60 @@ flag = true;
 % Calculate the loop time elapsed
 tic
 
+%% Conjugate Gradients algorithm
+
+% Calculate the gradient of 'f(x)' at 'x'
+g1 = df(x);
+
+% Calculate the decreasing direction
+d1 = -g1;
+
 % Conjugate Gradients algorithm
 while flag
-    % Calculate the gradient of 'f(x)' at 'x_old'
-    g_old = df(x_old);
-    d_old = -g_old;
 
-    % This costly 'alpha' is for finding the exact solution of the problem
-    % alpha = (-g_old' * d_old) / (d_old' * Q * d_old);
+    % Calling learning rate functions
+    % alpha = exact(Q, g1);
+    % alpha = armijo(alpha_max, c, f, x, d1, g1);
+    % alpha = wolfe(alpha_max, c1, c2, f, df, x, d1, g1);
+    % alpha = strong_wolfe(alpha_max, c1, c2, f, df, x, d1, g1);
+    % alpha = goldstein(alpha_max, c, f, x, d1, g1);
 
     % Find the new 'x'
-    x_new = x_old + alpha * d_old;
-    X(i+1, :) = x_new';
+    x = x + alpha * d1;
+    X(k+1, :) = x';
 
-    % Calculate the gradient of 'f(x)' at 'x_new'
-    g_new = df(x_new);
+    % Calculate the gradient of 'f(x)' at 'x'
+    g2 = df(x);
 
     % Calculate the decreasing direction
-    % beta = (g_new' * Q * d_old) / (d_old' * Q * d_old);
-    beta = (-g_new' * g_new) / (-g_old' * g_old);
-    d_new = -g_new + beta * d_old;
+    % beta = (g2' * Q * d1) / (d1' * Q * d1);
+    beta = (-g2' * g2) / (-g1' * g1);
+    d2 = -g2 + beta * d1;
     
-    % Using Euclidean norm-2 to check treshold
-    if norm(x_new - x_old, 2) <= treshold
+    % Using Euclidean norm-inf to check treshold
+    if norm(X(k+1, :)' - X(k, :)', inf) <= treshold
     % if x_new == zeros([n,1]) % Exact methods treshold
         flag = false;
     else
         % Update parameters
-        x_old = x_new;
-        g_old = g_new;
-        d_old = d_new;
+        g1 = g2;
+        d1 = d2;
     end
 
     % Update the iterator variable
-    i = i + 1;
+    k = k + 1;
 
     % Print the new 'x' using c-style
-    fprintf('x_%d = \n\n', i-1)
-    disp(x_new);
+    fprintf('x_%d = \n\n', k-1)
+    disp(x);
     fprintf('\n')
 end
+
+%% Output settings
 
 % Get the loop time elapsed
 t = toc;
 
 % Print the results using 'printer' function
-printer(f, Q, x, b, x_new, i, t);
+% norm_printer(f, A, x, b, k, t);
+quadratic_printer(f, x, k, t);