+++ newtons_method.m
@@ -1,5 +1,9 @@
+%% DESCRIPTION
+
 % NEWTON'S METHOD
 % Minimization of Unconstrained Quadratic problems
+
+%% Settings
 
 clc
 clear all
@@ -7,18 +11,20 @@ close all
 
 format long;
 
-% Generate new problem
-[n, Q, x, b, f, df, x_old] = quadratic_form();
+%% Problem
 
-% A matrix to store all 'x' vectors during iterations
-X = zeros(1, n);
-X(1, :) = x_old';
+% Call 'problem' function to generate a new problem
+[f, df, Q, x, X] = quadratic_problem();
 
-% Hyper parameters
+%% Hyper parameters
+
+% Set a treshold to stop the loop
 treshold = 1e-6;
 
-% Define an index for count the number of iterations 
-i = 1;
+%% Loop settings
+
+% Define an index for count the number of iterations
+k = 1;
 
 % Define a boolean flag variable to controle the while loop
 flag = true;
@@ -26,31 +32,32 @@ flag = true;
 % Calculate the loop time elapsed
 tic
 
-% Newton's Method algorithm
-while flag
-    % Find the new 'x'
-    x_new = x_old - (eye(size(Q)) / Q) * df(x_old);
-    X(i+1, :) = x_new';
+%% Newton's Method algorithm
 
-    % Using Euclidean norm-2 to check treshold
-    if norm(x_new - x_old, 2) <= treshold
+while flag
+    
+    % Find the new 'x'
+    x = x - (eye(size(Q)) / Q) * df(x);
+    X(k+1, :) = x';
+
+    % Using Euclidean norm-inf to check treshold
+    if norm(X(k+1, :)' - X(k, :)', inf) <= treshold
         flag = false;
-    else
-        % Update parameters
-        x_old = x_new;
     end
 
     % Update the iterator variable
-    i = i + 1;
+    k = k + 1;
 
     % Print the new 'x' using c-style
-    fprintf('x_%d = \n\n', i-1)
-    disp(x_new);
+    fprintf('x_%d = \n\n', k-1)
+    disp(x);
     fprintf('\n')
 end
+
+%% Output settings
 
 % Get the loop time elapsed
 t = toc;
 
 % Print the results using 'printer' function
-printer(f, Q, x, b, x_new, i, t);
+quadratic_printer(f, x, k, t);