+++ quadratic_problem.m
@@ -1,30 +1,43 @@
-function [n, Q, x, b, f, df, x_old] = quadratic_form()
-   % Problem statement
-    n = 2;
+%% A function to generate a new problem
+
+function [n, f, df, Q, x, X] = quadratic_problem()
+    % Getting dimension of the problem
+    n = input("Enter the problem size: ");
+
+    % The interval of elements
     l = 100;
     h = 200;
+
     % First method to generate a positive definite matrix
     Q = gallery('randcorr', n);
     Q = round((h - l) * Q + l);
+
     % Second method to generate a positive definite matrix
-    % Q = unifrnd(l, h, n, n)
+    % Q = unifrnd(l, h, n, n);
     % Q = randi([l, h], n);
     % Q = Q * Q';
     % Q = Q + eye(n) * eps(max(abs(eig(Q)))) * 10;
+    
     % Third method to generate a positive definite matrix
     % Q = randn(n);
     % Q = Q + Q';
     % Q = Q + n * eye(n);
     % Q = (Q * Q');
+    
+    % Check if A is a Positive Definite (PD) matrix
     chol(Q);
-    x = sym('x', [n, 1]);
+    
+    % Generate a random vector
     b = randi([l, h], [n, 1]);
     
     % Define the function 'f(x)' and gradient of 'f(x)' 
     f = @(x) 1/2 * x' * Q * x - b' * x;
     df = @(x) Q * x - b;
-    % df = @(x) jacobian(f, x)';
     
     % Initial guess
-    x_old = randi([l, h], [n, 1]); 
+    x = randi([l, h], [n, 1]); 
+
+    % A matrix to store all 'x' vectors during iterations
+    X = zeros(1, n);
+    X(1, :) = x';
 end