+++ steepest_descent.m
@@ -1,5 +1,9 @@
+%% DESCRIPTION
+
 % STEEPEST DESCENT
 % Minimization of Unconstrained Quadratic problems
+
+%% Settings
 
 clc
 clear all
@@ -7,60 +11,64 @@ close all
 
 format long;
 
-% Generate new problem
-[n, Q, x, b, f, df, x_old] = quadratic_form();
+%% Problem
 
-% A matrix to store all 'x' vectors during iterations
-X = zeros(1, n);
-X(1, :) = x_old';
+% Call 'problem' function to generate a new problem
+[f, df, Q, x, X] = quadratic_problem();
 
-% Hyper parameters
+%% Hyper parameters
+
+% Set a treshold to stop the loop
 treshold = 1e-6;
-alpha = 0.001;
 
-% Define an index for count the number of iterations 
-i = 1;
+% Learning rate
+alpha = 0.0001;
 
-% define a boolean flag variable to controle the while loop
+%% Loop settings
+
+% Define an index for count the number of iterations
+k = 1;
+
+% Define a boolean flag variable to controle the while loop
 flag = true;
 
 % Calculate the loop time elapsed
 tic
 
-% Steepest Descent algorithm
+%% Steepest Descent algorithm
+
 while flag
-    % Calculate the gradient of 'f(x)' at 'x_old'
-    g = df(x_old);
+    % Calculate the gradient of 'f(x)' at 'x'
+    g = df(x);
 
     % Calculate the decreasing direction
     d = -g;
-    
-    % This costly 'alpha' is for finding the exact solution of the problem
-    % alpha = (-g' * g) / (g' * Q * g);
+
+    % Calling learning rate functions
+    % alpha = exact(Q, g);
     
     % Find the new 'x'
-    x_new = x_old + alpha * d;
-    X(i+1, :) = x_new';
+    x = x + alpha * d;
+    X(k+1, :) = x';
 
-    % Using Euclidean norm-2 to check treshold
-    if norm(x_new - x_old, 2) <= treshold
+    % Using Euclidean norm-inf to check treshold
+    if norm(X(k+1, :)' - X(k, :)', inf) <= treshold
         flag = false;
-    else
-        % Update parameters
-        x_old = x_new;
     end
 
     % Update the iterator variable
-    i = i + 1;
+    k = k + 1;
 
     % Print the new 'x' using c-style
-    fprintf('x_%d = \n\n', i-1)
-    disp(x_new);
+    fprintf('x_%d = \n\n', k-1)
+    disp(x);
     fprintf('\n')
 end
+
+%% Output settings
 
 % Get the loop time elapsed
 t = toc;
 
 % Print the results using 'printer' function
-printer(f, Q, x, b, x_new, i, t);
+quadratic_printer(f, x, k, t);